requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.4.2
python-multipart>=0.0.6
faiss-cpu>=1.7.4
langchain_faiss>=0.0.1

# Authentication
PyJWT>=2.8.0

# LangChain and related packages
langchain>=0.1.0
langchain_core>=0.1.0
langchain_text_splitters>=0.0.1
langchain_google_genai>=0.0.5
langchain_openai>=0.0.5

# Google Gemini
google-generativeai>=0.3.0

# OpenAI
openai>=1.0.0

# PDF processing
pypdf>=3.17.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# Image processing
Pillow>=10.0.0

# Utilities
numpy>=1.24.0
requests>=2.31.0


